+++ pages/new/index.tsx
@@ -8,7 +8,12 @@ import { useEffect } from "react";
 import { Heading } from "@nypl/design-system-react-components";
 import RoutingLinks from "../../src/components/RoutingLinks.tsx";
 import useFormDataContext from "../../src/context/FormDataContext";
-import { getCsrfToken } from "../../src/utils/utils";
+import {
+  generateNewCookieTokenAndHash,
+  generateNewToken,
+  parseTokenFromPostRequestCookies,
+} from "../../src/utils/csrfUtils";
+import * as cookie from "../../src/utils/CookieUtils";
 
 import { GetServerSideProps } from "next";
 import { useTranslation } from "next-i18next";
@@ -61,12 +66,24 @@ function HomePage({ policyType, csrfToken, lang }: HomePageProps) {
 }
 
 export const getServerSideProps: GetServerSideProps = async (context) => {
-  const { csrfToken } = getCsrfToken(context.req, context.res);
+  const tokenFromRequestCookie = parseTokenFromPostRequestCookies(context.req);
   const { query } = context;
-  context.res.setHeader(
-    "Set-Cookie",
-    `nyplUserRegistered=false; Max-Age=-1; path=/; domain=${cookieDomain};`
-  );
+  let newTokenCookie;
+  let csrfToken = tokenFromRequestCookie.value;
+
+  if (!csrfToken) {
+    csrfToken = generateNewToken();
+    const newTokenCookieString = generateNewCookieTokenAndHash(csrfToken);
+    newTokenCookie = cookie.buildCookieHeader(newTokenCookieString);
+  }
+
+  const headers = [
+    // reset cookie that would otherwise bump users out of the flow
+    // to succcess page
+    `nyplUserRegistered=false; Max-Age=-1; path=/; domain=${cookieDomain};`,
+    newTokenCookie,
+  ];
+  context.res.setHeader("set-cookie", headers);
 
   return {
     props: {